refactor: refactor users page components (#33)
close #33
