+++ src/app/users/components/Sections/index.ts
@@ -1,0 +1,5 @@
+export * from './ActionsSection'
+export * from './FiltersSection'
+export * from './LatestSearchesSection'
+export * from './ResultsSection'
+
