+++ src/app/users/components/index.ts
@@ -1,5 +1,6 @@
+export * from './AddUserModal/AddUserModal'
 export * from './FilterByCategory'
+export * from './Sections'
 export * from './TextFieldUser'
 export * from './User'
-export * from './UsersSection'
