+++ src/app/users/page.tsx
@@ -1,9 +1,14 @@
-import { UsersSection } from './components'
+import { ActionsSection, FiltersSection, LatestSearchesSection, ResultsSection } from './components'
 
 const DashboardPage = () => {
   return (
-    <div className='h-full'>
-      <UsersSection />
+    <div className='flex flex-col gap-3 h-full'>
+      <div className='flex-shrink-0 flex gap-3'>
+        <FiltersSection />
+        <ActionsSection />
+      </div>
+      <ResultsSection />
+      <LatestSearchesSection />
     </div>
   )
 }
