+++ src/components/index.ts
@@ -1,4 +1,3 @@
-export * from './AddUserModal/AddUserModal'
 export * from './CarouselImages'
 export * from './ui'
